Pass all logic into useCurrencyField hook
