+++ src/components/Other/CurrencyField/CurrencyField.tsx
@@ -2,7 +2,6 @@ import { Field } from 'formik';
 import { InputForm } from '../../../styles';
 import { TypeOfRecord } from '../../../globalInterface';
 import { CurrencyAdornment } from '../../UI/Records/features/TransactionFormFields/CurrencyAdornment';
-import { formatCurrencyToString, formatValueToCurrency } from '../../../utils';
 import { useCurrencyField } from './useCurrencyField';
 
 interface CurrencyFieldProps {
@@ -11,66 +10,11 @@ interface CurrencyFieldProps {
   setFieldValue: (name: string, value: string) => void;
   updateAmount: (amount: string) => void;
 }
-const CURRENCY_FIELD_NAME = 'amount';
 
 const CurrencyField = ({
   setFieldValue, updateAmount, amount, typeOfRecord,
 }: CurrencyFieldProps) => {
-  const { validateCurrencyField } = useCurrencyField();
-
-  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const newValue = event.target.value;
-    const hasDeletedCharacter = amount.length > newValue.length;
-    const hasDeletedNumber = /,\d{2}$/;
-    const hasNumericPeriodComma = /[0-9.,]+/;
-    const valueEndsWithPeriod = /[0-9]+[.]$/;
-    const isPeriod = /\./;
-    const hasComma = /,/;
-    if (!newValue && amount) {
-      updateAmount('0');
-      setFieldValue(CURRENCY_FIELD_NAME, '');
-      return;
-    }
-    // If the value is not a number, do not update the field
-    if (!newValue.match(hasNumericPeriodComma)) return;
-    // If the new character is a period and there is no number, do not update the field
-    if (amount === '' && newValue.match(isPeriod)) return;
-
-    if (hasDeletedCharacter && newValue.match(hasDeletedNumber)) {
-      const newAmountNotFormatted = formatCurrencyToString(newValue);
-      updateAmount(newAmountNotFormatted);
-      const newAmount = formatValueToCurrency({ amount: newAmountNotFormatted, hasNoDecimals: true, hasNoCurrencySign: true });
-      setFieldValue(CURRENCY_FIELD_NAME, newAmount);
-      return;
-    }
-
-    // If the new value is a thousand, formate the number
-    if (newValue.length === 4 && !newValue.match(hasComma)) {
-      if (newValue.match(valueEndsWithPeriod)) {
-        setFieldValue(CURRENCY_FIELD_NAME, newValue);
-        return;
-      }
-      updateAmount(newValue);
-      const newAmount = formatValueToCurrency({ amount: newValue, hasNoDecimals: true, hasNoCurrencySign: true });
-      setFieldValue(CURRENCY_FIELD_NAME, newAmount);
-      return;
-    }
-    if (newValue.length > 5 && newValue.match(hasComma)) {
-      if (newValue.match(valueEndsWithPeriod)) {
-        setFieldValue(CURRENCY_FIELD_NAME, newValue);
-        return;
-      }
-      const newAmountNotFormatted = formatCurrencyToString(newValue);
-      updateAmount(newAmountNotFormatted);
-      const newAmount = formatValueToCurrency({ amount: newAmountNotFormatted, hasNoDecimals: true, hasNoCurrencySign: true });
-      setFieldValue(CURRENCY_FIELD_NAME, newAmount);
-      return;
-    }
-    // Guardar el nuevo amount como string
-    // Transformo el monto a currency
-    updateAmount(newValue);
-    setFieldValue(CURRENCY_FIELD_NAME, newValue);
-  };
+  const { validateCurrencyField, handleChange } = useCurrencyField({ setFieldValue, amount, updateAmount });
 
   return (
     <Field
